backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
argparse==1.4.0           # via -r requirements.in
atomicwrites==1.4.0       # via pathlib-mate
attrs==20.3.0             # via pytest, uszipcode
autopep8==1.5.4           # via pathlib-mate
babel==2.9.0              # via flask-babelex
bcrypt==3.2.0             # via flask-bcrypt
blinker==1.4              # via flask-mail, flask-principal
cerberus==1.3.2           # via -r requirements.in
certifi==2020.12.5        # via requests, sentry-sdk
cffi==1.14.4              # via bcrypt
chardet==3.0.4            # via requests
click==7.1.2              # via flask, pip-tools, rq
croniter==0.3.36          # via rq-scheduler
dnspython==2.0.0          # via email-validator
ecdsa==0.14.1             # via python-jose
email-validator==1.1.2    # via flask-security-too, wtforms
flask-admin==1.5.7        # via -r requirements.in
flask-babelex==0.9.4      # via flask-security-too
flask-bcrypt==0.7.1       # via -r requirements.in
flask-cors==3.0.9         # via -r requirements.in
flask-jwt-extended==3.25.0  # via -r requirements.in
flask-login==0.5.0        # via flask-security-too
flask-mail==0.9.1         # via flask-security-too
flask-mongoengine==1.0.0  # via -r requirements.in
flask-principal==0.4.0    # via flask-security-too
flask-security-too==3.4.4  # via -r requirements.in
flask-wtf==0.14.3         # via flask-mongoengine, flask-security-too
flask==1.1.2              # via -r requirements.in, flask-admin, flask-babelex, flask-bcrypt, flask-cors, flask-jwt-extended, flask-login, flask-mail, flask-mongoengine, flask-principal, flask-security-too, flask-wtf
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via email-validator, requests
importlib-metadata==3.3.0  # via pluggy, pytest
iniconfig==1.1.1          # via pytest
itsdangerous==1.1.0       # via flask, flask-security-too, flask-wtf
jinja2==2.11.2            # via flask, flask-babelex
lob==4.0.2                # via -r requirements.in
markupsafe==1.1.1         # via jinja2, wtforms
mongoengine==0.21.0       # via flask-mongoengine
natsort==7.1.0            # via croniter
packaging==20.8           # via pytest
passlib==1.7.4            # via flask-security-too
pathlib-mate==1.0.1       # via uszipcode
phonenumbers==8.12.14     # via -r requirements.in
pip-tools==5.4.0          # via -r requirements.in
pluggy==0.13.1            # via pytest
py==1.10.0                # via pytest
pyasn1==0.4.8             # via python-jose, rsa
pycodestyle==2.6.0        # via autopep8
pycparser==2.20           # via cffi
pyjwt==1.7.1              # via flask-jwt-extended
pymongo==3.11.2           # via mongoengine
pyparsing==2.4.7          # via packaging
pytest==6.2.0             # via -r requirements.in
python-dateutil==2.8.1    # via -r requirements.in, croniter
python-dotenv==0.15.0     # via -r requirements.in
python-jose==3.2.0        # via -r requirements.in
pytz==2020.4              # via babel
redis==3.5.3              # via -r requirements.in, rq
requests==2.25.0          # via -r requirements.in, lob, trycourier, uszipcode
rq-scheduler==0.10.0      # via -r requirements.in
rq==1.7.0                 # via -r requirements.in, rq-scheduler
rsa==4.6                  # via python-jose
sentry-sdk==0.19.5        # via -r requirements.in
six==1.15.0               # via bcrypt, ecdsa, flask-cors, flask-jwt-extended, pathlib-mate, pip-tools, python-dateutil, python-jose
speaklater==1.3           # via flask-babelex
sqlalchemy==1.3.20        # via uszipcode
toml==0.10.2              # via autopep8, pytest
trycourier==1.6.0         # via -r requirements.in
typing-extensions==3.7.4.3  # via importlib-metadata
urllib3==1.26.2           # via requests, sentry-sdk
uszipcode==0.2.4          # via -r requirements.in
werkzeug==1.0.1           # via flask, flask-jwt-extended
wtforms[email]==2.3.3     # via flask-admin, flask-mongoengine, flask-wtf
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
